debug

+++ pyramid-mos/pyramid_train.py
@@ -70,7 +70,7 @@ for ep in range(1, epochs+1):
         lens = batch['lens']
         scores = batch['score'].to(device).unsqueeze(-1)
         pred_scores = model(aud, lens)
-        batch_loss = criterion(pred_scores, scores)
+        batch_loss = criterion(pred_scores, scores).float()
         optimizer.zero_grad()
         batch_loss.backward()
         optimizer.step()
